refactor: moved the code related to placing the user on the map to a new file
